requirements.txt
pyparsing==2.4.7
python-docx==0.8.10
docxtpl==0.11.3
Jinja2==2.11.3
lxml==4.6.2
httplib2==0.19.0
MarkupSafe==1.1.1

cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
idna==2.10
packaging==20.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.25.1
rsa==4.7
uritemplate==3.0.1
urllib3==1.26.3

oauth2client==4.1.3

